let topic mentionable
